Remove all but but grunt nightwatch tasks
