--- app/config.js
@@ -1,18 +1,0 @@
-require.config({
-    baseUrl: '.',
-    paths: {
-        'astro-base': '../node_modules/astro-sdk/js/src/astro-base',
-        'astro-full': '../node_modules/astro-sdk/js/src/astro-full',
-        'astro-rpc': '../node_modules/astro-sdk/js/src/global/astro-rpc',
-        'app-events': 'global/app-events',
-        'app-rpc': 'global/app-rpc',
-        'application': '../node_modules/astro-sdk/js/src/application',
-        'bluebird': '../node_modules/bluebird/js/browser/bluebird',
-        'config': 'app-config',
-        'plugins': '../node_modules/astro-sdk/js/src/plugins',
-        'controllers': '../node_modules/astro-sdk/js/src/controllers',
-        'plugin-manager': '../node_modules/astro-sdk/js/src/plugin-manager',
-        'vendor/backbone-events': '../node_modules/astro-sdk/js/vendor/backbone-events',
-        'worker': '../node_modules/astro-sdk/js/src/worker'
-    }
-});
