--- app/tasks/config/requirejs.js
@@ -1,15 +1,0 @@
-module.exports = function(grunt) {
-    return {
-        worker: {
-            options: {
-                almond: true,
-                wrap: true,
-                mainConfigFile: "./config.js",
-                optimize: "none",
-                keepBuildDir: true,
-                name: "app.js",
-                out: "./build/app.js"
-            }
-        }
-    };
-};
