--- app/tasks/config/watch.js
@@ -1,8 +1,0 @@
-module.exports = function(grunt) {
-    return {
-        default: {
-            files: ['*.js', 'vendor/*.js'],
-            tasks: ['requirejs:worker']
-        }
-    };
-};
