--- app/tasks/eslint/tasks.js
@@ -1,3 +1,0 @@
-module.exports = function(grunt) {
-    grunt.loadNpmTasks('grunt-eslint');
-};
