--- app/tasks/jslinting.js
@@ -1,6 +1,0 @@
-module.exports = {
-    targets: [
-        'tests/system/**/*.js',
-        '*.js'
-    ]
-};
