--- app/tasks/lint/tasks.js
@@ -1,5 +1,0 @@
-module.exports = function(grunt) {
-    grunt.registerTask('lint:dev', ['eslint:dev']);
-    grunt.registerTask('lint:prod', ['eslint:prod']);
-    grunt.registerTask('lint', ['lint:dev']);
-};
