+++ package.json
@@ -15,8 +15,6 @@
     "eslint-loader": "^1.6.1",
     "grunt": "^0.4.5",
     "grunt-cli": "^1.1.0",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-eslint": "^19.0.0",
     "mobify-code-style": "^2.2.1",
     "nightwatch": "0.9.6",
     "nightwatch-commands": "1.7.0",
